requirements.txt
fastapi
uvicorn
python-dotenv
ultralytics
opencv-python-headless==4.7.0.72
pillow
python-multipart